Minor: fix unit tests
