+++ tests/unit/test_cli_workflows.py
@@ -23,11 +23,13 @@ def test_workflows_command_lists_workflows():
     asyncio.run(repo.create_workflow(corr2, {"itinerary": []}, {}))
 
     runner = CliRunner()
-    result = runner.invoke(app, ["workflows"])
-    assert result.exit_code == 0
+    result = runner.invoke(app, ["workflow", "list"])
+    assert (
+        result.exit_code == 0
+    ), f"Command failed with exit code {result.exit_code}. Output: {result.stderr}"
     output = result.stdout
-    assert corr1 in output
-    assert corr2 in output
+    assert corr1 in output, f"Correlation ID {corr1} not found in output: {output}"
+    assert corr2 in output, f"Correlation ID {corr2} not found in output: {output}"
 
 
 def test_workflow_command_shows_details_and_missing():
@@ -38,13 +40,19 @@ def test_workflow_command_shows_details_and_missing():
     asyncio.run(repo.mark_step_completed(corr, "step1", status="completed"))
 
     runner = CliRunner()
-    result = runner.invoke(app, ["workflow", corr])
-    assert result.exit_code == 0
+    result = runner.invoke(app, ["workflow", "show", corr])
+    assert (
+        result.exit_code == 0
+    ), f"Command failed with exit code {result.exit_code}. Output: {result.stderr}"
     output = result.stdout
-    assert corr in output
-    assert "step1" in output
-    assert "completed" in output
+    assert corr in output, f"Correlation ID {corr} not found in output: {output}"
+    assert "step1" in output, f"Step 'step1' not found in output: {output}"
+    assert "completed" in output, f"Status 'completed' not found in output: {output}"
 
-    result_missing = runner.invoke(app, ["workflow", "missing-id"])
-    assert result_missing.exit_code == 1
-    assert "Workflow not found" in result_missing.stdout
+    result_missing = runner.invoke(app, ["workflow", "show", "missing-id"])
+    assert (
+        result_missing.exit_code == 1
+    ), f"Expected exit code 1 for missing workflow, got {result_missing.exit_code}. Output: {result_missing.stdout}"
+    assert (
+        "Workflow not found" in result_missing.stdout
+    ), f"Expected 'Workflow not found' message, got: {result_missing.stderr}"
